requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.3
django-environ==0.11.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
pycparser==2.22
sqlparse==0.4.4
tzdata==2024.1
